test(workspace): split app center service tests
